fix: add 2and3 validator for path parameter declaration validation
